Auth

+++ src/firebase/deleteDocument.js
@@ -1,8 +1,22 @@
 import { deleteDoc, doc } from 'firebase/firestore';
-import { projectFirestore } from './config';
+import { projectFirestore, auth } from './config';
 
-const deleteDocument = (collectionName, documentId) => {
-  return deleteDoc(doc(projectFirestore, collectionName, documentId));
+
+const deleteDocument = async (collectionName, documentId) => {
+  const user = auth.currentUser;
+  if (!user) {
+    alert('You must be logged in to delete this document.');
+    return;
+  }
+
+  const idTokenResult = await user.getIdTokenResult();
+  if (!idTokenResult.claims.admin) {
+    alert('You are not authorized to delete this document.');
+    return;
+  }
+
+  await deleteDoc(doc(projectFirestore, collectionName, documentId));
+  alert('Document deleted successfully.');
 };
 
 export default deleteDocument;
